\name{TeachingPopGen}
\alias{TeachingPopGen}
\docType{package}
\title{Teaching Population Genetics}
\description{
This package \pkg{TeachingPopGen} provides a number of basic functions for analysis of population genetic data. For a list of help files on functions, as well as topical vignettes, type \code{library(help=TeachingPopGen)} or press on \code{Index} below.
}
\keyword{package}
